refactor to improve maintainability
